Added a lot to the about.jsp, added blog/welcome servlets and jsps, renames homeservlet to newsengineservlet, added instructions to the newsengine
