+++ scitweets/src/main/java/adminControls/AdminLoginServlet.java
@@ -8,7 +8,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import Properties.RetrieveProperties;
+import properties.RetrieveProperties;
 
 @SuppressWarnings("serial")
 @WebServlet(urlPatterns = "/adminlogin")
