+++ scitweets/src/main/java/dbconnection/DBConnect.java
@@ -10,7 +10,7 @@ import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 
-import Properties.RetrieveProperties;
+import properties.RetrieveProperties;
 
 public class DBConnect {
