+++ scitweets/src/main/java/welcome/WelcomeServlet.java
@@ -1,0 +1,17 @@
+package welcome;
+
+import java.io.IOException;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@SuppressWarnings("serial")
+@WebServlet(urlPatterns = "/home")
+public class WelcomeServlet extends HttpServlet{
+	protected void doGet(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
+		request.getRequestDispatcher("/WEB-INF/views/welcome.jsp").forward(request, response);
+	}
+}
