+++ scitweets/src/main/webapp/WEB-INF/views/blog.jsp
@@ -1,0 +1,14 @@
+<%@ page language="java" contentType="text/html; charset=ISO-8859-1"
+    pageEncoding="ISO-8859-1"%>
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+<html>
+<head>
+<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
+<title>Scitweets: Keep Current With the Scientific World</title>
+</head>
+<body>
+	Blogs are currently under construction. Keep checking back for updates!
+	<br>
+	<a href="/home">Click here to return home</a>
+</body>
+</html>
